+++ bisonc++/documentation/manual/algorithm/examples/Parserbase.h
@@ -27,7 +27,7 @@ typedef CHARP STYPE;
 
     private:
         int d_stackIdx;
-        std::vector<unsigned>   d_stateStack;
+        std::vector<size_t/*unsigned*/>   d_stateStack;
         std::vector<STYPE>      d_valueStack;
 
     protected:
@@ -42,9 +42,9 @@ typedef CHARP STYPE;
             UNEXPECTED_TOKEN,
         };
         bool        d_debug;
-        unsigned    d_nErrors;
+        size_t/*unsigned*/    d_nErrors;
         int         d_token;
-        unsigned    d_state;
+        size_t/*unsigned*/    d_state;
         STYPE      *d_vsp;
         STYPE       d_val;
 
@@ -59,14 +59,14 @@ typedef CHARP STYPE;
         {
             return d_debug;
         }
-        void pop(unsigned count = 1);
-        void push(unsigned nextState);
-        unsigned reduce(PI const &productionInfo);
+        void pop(size_t/*unsigned*/ count = 1);
+        void push(size_t/*unsigned*/ nextState);
+        size_t/*unsigned*/ reduce(PI const &productionInfo);
         void setDebug(bool mode)
         {
             d_debug = mode;
         }
-        unsigned top() const;
+        size_t/*unsigned*/ top() const;
 
 // class ParserBase ends
 };
