+++ bisonc++/documentation/manual/examples/mfcalc/parser/Parserbase.h
@@ -37,7 +37,7 @@ struct STYPE
 
     private:
         int d_stackIdx;
-        std::vector<unsigned>   d_stateStack;
+        std::vector<size_t/*unsigned*/>   d_stateStack;
         std::vector<STYPE>      d_valueStack;
 
     protected:
@@ -52,9 +52,9 @@ struct STYPE
             UNEXPECTED_TOKEN,
         };
         bool        d_debug;
-        unsigned    d_nErrors;
+        size_t/*unsigned*/    d_nErrors;
         int         d_token;
-        unsigned    d_state;
+        size_t/*unsigned*/    d_state;
         STYPE      *d_vsp;
         STYPE       d_val;
 
@@ -69,14 +69,14 @@ struct STYPE
         {
             return d_debug;
         }
-        void pop(unsigned count = 1);
-        void push(unsigned nextState);
-        unsigned reduce(PI const &productionInfo);
+        void pop(size_t/*unsigned*/ count = 1);
+        void push(size_t/*unsigned*/ nextState);
+        size_t/*unsigned*/ reduce(PI const &productionInfo);
         void setDebug(bool mode)
         {
             d_debug = mode;
         }
-        unsigned top() const;
+        size_t/*unsigned*/ top() const;
 
 // class ParserBase ends
 };
