+++ bisonc++/grammar/reduce.cc
@@ -4,7 +4,7 @@ void Grammar::reduce(StateInfo &state)
 {
     d_token = d_stateInfo[d_state].lhs();
 
-    for (unsigned idx = state.reductionSize(); idx--; )
+    for (size_t/*unsigned*/ idx = state.reductionSize(); idx--; )
             d_stack.pop_back();
 
     d_state = d_stack.back();
