+++ bisonc++/item/item2.cc
@@ -1,6 +1,6 @@
 #include "item.ih"
 
-Item::Item(Item const *item, unsigned dot)
+Item::Item(Item const *item, size_t/*unsigned*/ dot)
 :
     d_production(item->d_production),
     d_dot(dot)
