+++ bisonc++/item/item3.cc
@@ -1,6 +1,6 @@
 #include "item.ih"
 
-Item::Item(Production const *production, unsigned dot)
+Item::Item(Production const *production, size_t/*unsigned*/ dot)
 :
     d_production(production),
     d_dot(dot)
