+++ bisonc++/nonterminal/data.cc
@@ -1,4 +1,4 @@
 #include "nonterminal.ih"
 
-unsigned NonTerminal::s_counter;
-unsigned NonTerminal::s_number;
+size_t/*unsigned*/ NonTerminal::s_counter;
+size_t/*unsigned*/ NonTerminal::s_number;
