+++ bisonc++/production/vectoridx.cc
@@ -1,6 +1,6 @@
 #include "production.ih"
 
-Symbol *Production::vectorIdx(unsigned idx) const
+Symbol *Production::vectorIdx(size_t/*unsigned*/ idx) const
 {
     return idx >= size() ? 0 : std::vector<Symbol *>::operator[](idx);
 }
